+++ src/components/Column.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ColumnItem from '../components/ColumnItem';
+import ColumnTitle from './ColumnTitle';
+import NewTaskInput from './NewTaskInput';
 
 class Column extends React.Component {
   render() {
@@ -43,68 +45,10 @@ class Column extends React.Component {
   }
 }
 
-class ColumnTitle extends React.Component {
-  constructor(props) {
-    super(props);
-    this.focused = React.createRef();
-  }
-  componentDidUpdate() {
-    if (this.focused.current) {
-      this.focused.current.focus();
-    }
-  }
-  render() {
-    const {isTitleEditable, columnNumber, onClickColumnTitle, editColumnTitle, columnTitle} = this.props;
-    return isTitleEditable 
-      ? 
-      <form onSubmit={() => onClickColumnTitle(columnNumber)}>
-        <input className="column__title"
-          onBlur={() => onClickColumnTitle(columnNumber)}
-          onChange={event => editColumnTitle(columnNumber, event.target.value)}
-          value={columnTitle}
-          ref={this.focused}
-        />
-      </form>
-      : 
-      <span className="column__title" onClick={() => onClickColumnTitle(columnNumber)}>
-        {columnTitle}
-      </span>
-      
-  }
-}
-
-const NewTaskInput = ({columnNumber, inputValue, onInputTask, onSubmitTask}) => {
-  const submitTask = (e, columnNumber, inputValue) => {
-    e.preventDefault();
-    onSubmitTask(columnNumber, inputValue);
-  };
-  return (
-    <form className="column__form" onSubmit={() => submitTask(columnNumber, inputValue)}>
-      <input className="column__input" value={inputValue} onChange={event => onInputTask(columnNumber, event.target.value)} />
-      <button className="column__submit-button" type="submit">+</button>
-    </form>
-  )
-}
-
 Column.propTypes = {
   columnNumber: PropTypes.number.isRequired,
   columnState: PropTypes.object.isRequired,
   actions: PropTypes.object.isRequired,
-};
-
-ColumnTitle.propTypes = {
-  isTitleEditable: PropTypes.bool.isRequired,
-  columnNumber: PropTypes.number.isRequired,
-  onClickColumnTitle: PropTypes.func.isRequired,
-  editColumnTitle: PropTypes.func.isRequired,
-  columnTitle: PropTypes.string.isRequired,
-};
-
-NewTaskInput.propTypes = {
-  columnNumber: PropTypes.number.isRequired,
-  inputValue: PropTypes.string.isRequired,
-  onInputTask: PropTypes.func.isRequired,
-  onSubmitTask: PropTypes.func.isRequired,
 };
 
 export default Column;
